dev_requirements.txt
alabaster==0.7.12
amqp==2.5.2
apipkg==1.5
apns2==0.7.1
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bbcode==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coverage==5.0.4
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==2.2
    django-admin-charts==0.19.0
    django-admin-honeypot==1.1.0
    django-admin-tools==0.9.0
    django-annoying==0.10.6
    django-appconf==1.0.4
    django-bootstrap-form==3.4
    django-bower==5.2.0
    django-check-seo==0.3.4
    django-ckeditor==5.9.0
    django-classy-tags==1.0.0
    django-cms==3.7.2
django-cors-middleware==1.5.0
django-debug-toolbar==2.2
django-defender==0.7.0
django-downtime==1.1.2
django-environ==0.4.5
django-extensions==2.2.9
django-extra-views==0.13.0
django-feedback==1.1.1
django-filer==1.7.1
django-formtools==2.2
django-haystack==3.0b1
django-helpdesk==0.2.21
django-jet==1.0.8
django-js-asset==1.2.2
django-jsonfield==1.4.0
django-knowledge==0.2.0
django-machina==1.1
django-markdown-deux==1.0.5
django-markup-deprecated==0.0.3
django-meta==1.6.1
django-mobile==0.7.0
django-mobileadmin==0.5.2
django-modelcluster==5.0.1
django-mptt==0.11.0
django-newsletter==0.9
django-nvd3==0.9.7
django-oauth-toolkit==1.3.2
-e git+https://github.com/django-oscar/django-oscar.git@2429ad9e88e9a432dfa60aaca703d99860f85389#egg=django_oscar
django-oscar-accounts==2.0.0
django-oscar-api==2.0.2
django-oscar-api-checkout==1.0.0
django-oscar-bluelight==1.0.0
django-oscar-invoices==0.2
django-oscar-promotions==1.0.0b1
django-oscar-wagtail==0.2.0
django-otp==0.9.0
django-otp-yubikey==0.5.2
django-phonenumber-field==3.0.1
django-polymorphic==2.1.2
django-post-office==3.4.0
django-progressive-web-app==0.1.1
django-prometheus==2.0.0
django-push-notifications==2.0.0
django-pwa-webpush==1.1.0
django-redis==4.11.0
django-responsive2==0.1.3
django-rest==0.0.1
django-robots==4.0
django-sekizai==1.1.0
django-simple-history==2.10.0
django-smuggler==0.9.1
django-social-share==1.4.0
django-sphinxdoc==2.0.0
django-tables2==2.2.1
django-taggit==1.2.0
django-tellme==0.6.5
django-treebeard==4.3.1
django-two-factor-auth==1.11.0
django-user-agents==0.4.0
django-user-tracking==0.4.11
django-uwsgi==0.2.2
django-webtest==1.9.7
django-widget-tweaks==1.4.8
djangocms-admin-style==1.5.0
djangocms-page-meta==0.10.0
djangorestframework==3.11.0
djcacheutils==3.0.0
docutils==0.16
draftjs-exporter==2.1.7
easy-thumbnails==2.7
email-reply-parser==0.5.9
entrypoints==0.3
execnet==1.7.1
fabric==2.5.0
factory-boy==2.12.0
Faker==4.0.3
filelock==3.0.12
flake8==3.7.9
flake8-blind-except==0.1.1
flake8-debugger==3.2.1
freezegun==0.3.15
h2==2.6.2
hpack==3.0.0
html5lib==1.0.1
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.0
importlib-resources==1.4.0
instrumented-soap==2.0.0.post1
invoke==1.4.1
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jsonfield==3.1.0
kombu==4.6.8
l18n==2018.5
ldif3==3.1.1
lxml==4.5.0
Markdown==3.2.1
markdown2==2.3.8
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
paramiko==2.7.1
parso==0.7.0
pdfkit==0.6.1
pexpect==4.8.0
phonenumbers==8.12.2
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
purl==1.5
py==1.8.1
py-vapid==1.7.0
pybbm==0.19.0
pycodestyle==2.5.0
pycountry==19.8.18
pycparser==2.20
pycryptodome==3.9.7
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyprof2calltree==1.4.4
pysolr==3.8.1
pytest==5.4.1
pytest-django==3.7.0
pytest-forked==1.1.3
pytest-xdist==1.31.0
python-card-me==0.9.3
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.6.0
PyYAML==5.3.1
qrcode==6.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
soupsieve==2.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
suds-jurko==0.6
surlex==0.2.0
text-unidecode==1.3
toml==0.10.0
tox==3.14.6
traitlets==4.3.3
ua-parser==0.10.0
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.9
user-agents==2.1
uWSGI==2.0.18
vine==1.3.0
virtualenv==20.0.18
wagtail==2.9
waitress==1.4.3
wcwidth==0.1.9
webencodings==0.5.1
WebOb==1.8.6
WebTest==2.0.34
Werkzeug==1.0.1
whitenoise==5.0.1
Whoosh==2.7.4
Willow==1.3
XlsxWriter==1.2.8
YubiOTP==0.2.2.post1
zipp==3.1.0
